Use completion feedback to control read stream I/O.
Generally, if I/O is needed, we boost the look-ahead distance by
doubling it.  This produces both combined I/O and concurrent I/O.
It also means that we get up to higher look-ahead distances even when
there is no real benefit, wasting resources.

In preparation for real AIO, adjust the algorithm to continually adjust
the look-ahead distance so that all I/Os complete within small window.
It'll be boosted if the an I/O that is 1/8th of the remaining I/O queue
into the future would currently stall (though we haven't come to it
yet so there is still a chance it won't).  It will decay if an I/O that
is 2/8ths of the remanining I/O queue *would not* stall.  This
arrangement is intended to discover the smallest distance that avoids
stalls, ie without excessive I/O concurrency and buffer usage.

Without AIO, behavior degrades to always triggering distance increases
up to configured limits.