XXX: Window-based distance control for read_stream.c.
XXX Experimental idea #1: simple window
XXX This only tries to avoid excessive distance when I/O is already
keeping up with demand
XXX It doesn't try to avoid oversaturating the I/O subsystem
XXX It has numerous problems and is not very good!

This commit adds a stall warning window consisting of two sample I/Os
at the 25% and 50% points (picked because they're easy, not because
they're good!) in the current I/O queue, and checks whether they would
currently stall *if* we were to wait for them now (though we won't yet).
If the near sample would, it boosts the distance as before.  If it
wouldn't, but the far sample also wouldn't, then it begins to adjust the
distance downwards very slowly (see sustain counter, trying to reduce
sawtooth/oscillation, not entirely successfully).  If we don't have
enough depth for a window, it adjusts up more slowly, just to try to get
a window.

A later patch adds instrumentation to illustrate, something like:

  .                      --- boost on minimum I/O startup
  .[S                    --- boost on stall warning
  .[n s].                }-- hold on window = no-stall, stall
  .[n s].                }
  .[n s].                }
  .[n s].                }
  .[S .].                --- boost on stall warning
  . .[n . N]. . .        }-- decay on excessive distance indicator
  . .[n . N]. . .        }
  . .[n . N]. . .        }
  . .[n . s]. .          }-- hold on window = no-stall, stall
  . .[n . s]. .          }
  . .[n . s]. .          }