Handle hit/miss patterns better in read stream.
The previous algorithm would double ("attack") the look-ahead distance
for misses, and "decay" by one for each hit.  This didn't work too well
for a repeating hit-miss-hit-miss pattern, where we cloud clearly
benefit from I/O concurrency but we'd be stuck with a distance of 1.
Introduce a new counter "sustain", set to 16 when performing an I/O, and
decremented by hits.  Only when it is zero does the distance begin to
decay.  This helps boost us out of that pattern.

XXX Better ideas?
XXX Need forwarding patches too or the seq detector is confused by the
secret hit...