Fix OAUTH on NetBSD.
NetBSD's kqueue implementation doesn't like zero timeouts and fails with
EINVAL.  Steal the workaround from the same problem on Linux a few lines
up: use 1 instead, if libcurl asks for a timer in 0.

This could be seen on the optional NetBSD CI task, but not on the build
farm animals because they aren't finding curl and testing this code.

ci-os-only: openbsd netbsd freebsd