ci: Enable NetBSD and OpenBSD by default.
Previously you had to trigger them manually.  This will run them always,
since at least NetBSD is as fast as the others now, and OpenBSD still
beats Windows.

XXX Not saying we're quite ready to do this (are we?), I just want cfbot
to run them for this CF entry!