POC: Server support for TAP "wait_for_catchup".
Provide new built-in functions for use on a primary, to wait for replay
to be reported by a standby.  This is intended to be a drop-in
replacement for Cluster.pm's polling wait_for_catchup(), and is
retro-fitted.  The new fuctions are:

   pg_last_standby_lsn(standby_name, field_name)
   pg_wait_standby_lsn(standby_name, field_name, lsn, timeout_ms)

Field is "sent", "write", "flush", "replay".

The wait version waits for the named standby to connect (if it hasn't
already) and report a high enough LSN.  Once the timeout is reached, it
gives up and returns whatever it has, so the caller must compare it with
the expected LSN to distinguish timeout from success.  (XXX is that
weird?  Return NULL instedd?)

XXX POC