XXX: Synchronous WaitReadBuffersMightBlock().
Provide a basic implementation of WaitReadBuffersMightBlock() using
Linux 4.14's RWF_NOWAIT, and fall back to assuming that all I/Os might
block on other systems.

XXX This is only a toy implementation: it adds a check that we can read
the first byte of the block range with RWF_NOWAIT, as a cheap way to
estimate whether a whole I/O is already in the kernel's page cache.
This estimation technique is intended to check whether a recent
smgrprefetch(), the kernel's readahead or unrelated recent activity have
already faulted in the relevant region of the file, but of course it can
be misled by a partial fragment of the I/O range that happened to be in
its page cache; occasional mistakes like that are probably not terribly
consequential (?) for these heuristics.  You could probably invent an
adversarial test where it all falls down.  Shrug.

XXX This technique only makes economic sense if the caller is predicting
that it is *not* in the page cache, for example when asking whether the
far sample I/O would block.  The adaptive feedback algorithm tries to
arrange for that one not to be finished yet, if your storage is slow
enough for that to be possible.  It means we perform a total or two or
three system calls per I/O: sequential needs only two, this preadv2(1
byte) probe when the I/O is halfway through the I/O queue (ie to find
out if the kernel's readahead is keeping up with our consumer), and
preadv() when we actually need the data to give to the consumer; random
needs three, as above but also posix_fadvise() up front when the I/O
enters the queue.

XXX This toy implementation also does a preadv2(1 byte, RWF_NOWAIT) when
the I/O reaches the near sample, so add one to the above system call
count.  A more serious implementation could add a function
smgrreadv_nowait() for probing the near sample, since we expect that one
to succeed without waiting.  That is, go through the whole
BM_IO_IN_PROGRESS dance and then perform the whole preadv2(nblocks,
RFW_NOWAIT) with a hypothetical WaitReadBufferNoWait().  If it
succeeded, you could mark the I/O as done and then the later
WaitReadBuffers() call would become a no-op.  If it fails, then the
later WaitReadBuffers() call would do the whole BM_IO_IN_PROGRESS setup
again, but in that case the flow control algorithm would boost the
distance to avoid that in the future rather vigorously.  In other words,
there'd be two chances to read the data, once at the near sample point
in the queue, which we expect to succeed nearly 100% of the time, and
the maybe once again at buffer consumption time if we screwed up the
distance.

XXX I don't intend to implement that more seriously for now, since we're
on the cusp of having real AIO, but it's a viable way to demonstrate the
flow control algorithm even without AIO.  Probably should have tried
this earlier...  (First step for me would be to implement RWF_NOWAIT for
FreeBSD :-))

XXX It's remotely possible that even with AIO, some elements of this
thinking could be beneficial for buffered I/O.  It could help you avoid
wasting precious I/O workers on extra I/O depth when all they're doing
for you is copying data out of the kernel's cache: the memcpy()
concurrency is actually quite a nice benefit of io_method=worker, but a
shorter look-ahead distance can achieve that without pinning so many
buffers or launching an initial storm of I/Os in the stream ramp-up
phase.  Just something to think about...