aio: Add adaptive feedback I/O stream controller.
WORK IN PROGRESS

Previously read_stream.c would double the look-ahead distance every time
I/O was needed, limited only by *_io_concurrency and the allowed number
of buffers.  That minimizes stalling, but might overshoot considerably.

This commit adds a new stream controller that monitors the queue of I/Os
to find out when they have completed, so that it knows when the current
level of I/O concurrency is "enough" to satisfy the stream's current
rate of consumption.  This involves three steps:

1.  Choosing a target number of "done" I/Os to have ready at all times,
like the stock level in a shop.  The number should be high enough that
no stalls are expected considering recent fluctuations, and not more.
The model used assumes a normal distributions, and sets the target to
four standard deviations of the recently observed value.  Integer
arithmetic is used, except for a floating point sqrt() needed only when
the variance moves outside a band.

2.  Using a PID controller from control theory to drive the number of
done IOs to the target by adjusting the IO concurrency level.

3.  XXX VAPOURWARE: Defending against overloaded IO by detecting when it
is impossible to achieve the target, and starting more IOs is just
causing they to be queued up somewhere.

AioStreamController is separate from ReadStream, because it will be
needed by other kinds of streams, most obviously WriteStream.

Co-authored-by: TM
Co-authored-by: MP