XXX: read_stream.c I/O queue instrumentation
Throwaway instrumentation for read_stream.c flow control algorithm that
captures and then logs a history of the I/O queue including stall
detection window and distance adjustment decisions at DEBUG1 level.

Probably some elements of this would be useful in the tree, eg the basic
statistics counters, might be useful in EXPLAIN etc if we can figure out
where they should go.  (?)

If you apply the adaptive feedback patch but not the non-stub
WaitReadBuffersMightStall() patch, or you are on a non-Linux system (no
RWF_NOWAIT), then the output dumped by this patch reflects the behavior
of master without the adaptive feedback patch.  (Because the fallback
behavior matches master as of commit 799959dc.)