Deprecate pg_int64.
Previously we used pg_int64 in the function prototypes of libpq-fe.h
(introduce by commit 461ef73f).  It was a typedef for the type that we
selected for int64 in the backend.  That type is now int64_t (since
commit 962da900).

Given that it didn't match the name we use in the backend anyway, let's
just use int64_t directly in (rare) public interfaces.  It is likely to
be the name that modern applications use for their own 64-bit processing
so there is no need for yet another name.

Keep a typedef marked deprecated for backwards compatibility, but move
it to libpq-fe.h, where the only uses of it were.

Reviewed-by: Peter Eisentraut <peter@eisentraut.org>
Discussion: https://postgr.es/m/CA%2BhUKGKn_EkNNGMY5RzMcKP%2Ba6urT4JF%3DCPhw_zHtQwjvX6P2g%40mail.gmail.com