Add async-signal-safe ConditionVariableWakeOne().
In order to signal a CV from a signal handler in a later patch, add a
special-purpose wake function.  This has a very narrow intended usage
and slightly different semantics from the regular APIs, and is
clearly documented as such.  It also has a corresponding
ConditionVariableOrLatchSleep() function which is required when waiting
for ConditionVariableWakeOne() *or* the regular broadcast/signal
functions.

It has previously been mooted by bc971f40 that it should be possible to
include CVs in a WaitEventSet, which would obviate the need for the new
sleep function added here, but that seems like a more complex redesign
of the CV implementation.