aio: Improve I/O worker queue when full/contended.
Previously, when the submission queue was full we'd run the remaining
IOs in a whole batch synchronously.  Now we'll try again between
synchronous operations, because the queue might now have space.

XXX The follow part is a random experiment:

Commit 29a0fb21 made lock acquisition is conditional, with failure
treated the same way as a full queue.  Insert a couple of spin delay in
that case, to give the lock holder a chance of releasing it before we
try again.  Pin a tail on it and call it an adaptive mutex.

Discussion: https://postgr.es/m/CA%2BhUKG%2Bm4xV0LMoH2c%3DoRAdEXuCnh%2BtGBTWa7uFeFMGgTLAw%2BQ%40mail.gmail.com