Provide EventFlag IPC mechanism.
An EventFlag can be used to wait for something to happen, allowing
exactly one interested backend to process the event when it is signaled.
It is designed to be safe to signal from inside a Unix signal handler or
Windows thread pool, where (for example) LWLock and ConditionVariable
can't be used.  This is a building block that later patches for native
IO methods that don't allow completion notifications to be consumed
directly if the IO was submitted by another process.