aio: Add io_method=posix_aio.
Currently this works well only on FreeBSD, which has vectored
extensions.  It compiles and pass tests on any system with a
standard-conforming-enough POSIX AIO implementation, though note that
glibc just emulates POSIX AIO with user mode threads, and performs
terribly.

It uses realtime signals as the completion notification mechanism.
There are more efficient options including kqueue on FreeBSD, but they
aren't compatible with the current multi-process design of PostgreSQL,
so they might become more interesting in the future.  Unfortunately this
means that macOS can't run or even compile it, due to lack of realtime
signals.

io_method=posix_aio is enabled for the FreeBSD CI task.