aio: Add READ_STREAM_RESET_CONTINUE flag.
For users that sometimes run out of block numbers temporarily but then
later want to continue, provide a way to reset the stream and continue
without forgetting the internal look-ahead distance.

This is done as a bitmask, just in case we think of more reasons for
fine-grained control of what is reset or not in the future.

XXX Name debatable; is it better to use a 'high level' concept like
'continuing an interrupted stream', or a lower level concept explicitly
referring to what is reset or not?

XXX Where existing users pass 0 or the new flag not thought about very
hard

Discussion: https://postgr.es/m/CA%2BhUKG%2BWWr4-8TYemyU%3DucQsNe6bUBN_Sq3mCnBoBtxaJ9w3ug%40mail.gmail.com