Experiment:  Peeking ahead in I/O streams.
Allow consumers of self-referential block streams to install a callback
that gets a chance to "peek" at each block as soon as possible.

Consumers of self-referential block streams can use this to feed
navigational information to the block number callback much earlier in
the pipeline while they are still consuming older blocks.

This can't help if there is an IO stall in the way of progress at every
jump, but some mixed hit/miss patterns may be able to generate some
extra concurrency this way.

XXX highly experimental idea exploration...