Teach read_stream.c to combine blocks backwards.
Previously it could only build combined IOs when asked for blocks in
increasing order.  Now it can also detect reverse order scans.

This requires tracking regions of the buffer queue that need to be
returned in reverse order, so that it can wait for the whole range and
then stream them backwards.

XXX proof of concept