POSIX job control for child process cleanup.
When a session leader process exits, the foreground process group
receives SIGHUP.  Normally this kills the processes started
under a login shell when eg an ssh session disconnects.  We can set the
postmaster up the same way with some obscure but fully standardized
incantations.

SIGHUP is also still used to tell children to reload their
configuration file.  It would probably be better not to multiplex the
signal, but we can distinguish the cases by checking the postmaster
pipe.

We now have to use a single process group for all children so that they
can all be in the 'foreground'.  This fits with future plans for a
multi-threaded mode where there would be only one sub-process containing
all backends.

XXX But now SIGTERM isn't forwarded to grandchildren.