Support wctype.h classification for EUC encodings.
Instead of giving up on non-ASCII characters, convert pg_wchar values
using the PG_WCHAR_CUSTOM encoding scheme to wchar_t so that we can use
wctype.h functions on any character.

XXX This replaces the _ascii() version from the previous patch, to
experiment with a different approach

XXX Is this too expensive?