Replace pg_attribute_aligned() with C11 alignas().
Remove the compiler-specific macros and adopt the standard.

One use of pg_attribute_aligned() macro could not be replaced: it forced
the alignment of int128 and uint128, typedefs for signed and unsigned
__int128.  Since alignas() doesn't allow typedefs to have their own
alignment separate from the underlying type, we need another solution to
the problem it solved, namely that palloc() would not allocate memory
aligned for this type.  That stems from MAXIMUM_ALIGN not including
alignof(__int128).  The goal of explicit alignment was to prevent the
compiler from generating code that requires __int128's natural alignment
of 16, at the cost of losing SSE instructions for 128 bit arithmetic.

Since commit 751804998, we have gained palloc_aligned().  Use that
instead, for the struct that makes use of the type.