meson: Try to find libintl without -Dextra_XXX.
Since libintl is the only dependency that Meson can't usually find via
pkg-config on typical non-Linux Unix systems, let's try a bit harder to
find it with existing clues.

We typically find msgfmt in $PATH, so we can plausibly guess that the
library and headers are installed in the same prefix, and then try that
before giving up and requiring a manual path configuration.

This revealed some missing dependency declarations previously covered by
the catch-all extra_includes_dir, added here.

XXX Is this a terrible idea for some reason?