Refactor static_assert() support.
In meson, HAVE__STATIC_ASSERT was really a test for GCC
statement expressions, as needed for StaticAssertExpr() now that
_Static_assert could be assumed to be available through our C11
requirement.  This artificially prevented Visual Studio from being able
to use static_assert() in other contexts.

The configure script only tested for _Static_assert, which was
tautological (we require C11), but also incorrectly triggered the
assumption about statement expressions in c.h.

Instead, make a new test for HAVE_STATEMENT_EXPRESSIONS, and use that
to control only whether StaticAssertExpr() uses fallback code, not the
other variants.  This improves the quality of failure messages in the
(much more common) other variants under Visual Studio.

Also get rid of the two separate implementations for C++, since the C
implementation is also also valid as C++11.  While it is a stretch to
apply HAVE_STATEMENT_EXPRESSIONS tested with $CC to a C++ compiler, the
previous C++ coding assumed that the C++ compiler had them
unconditionally, so it isn't a new stretch.  In practice, the C and C++
compilers are very likely to agree, and if a combination is ever
reported that falsifies this assumption we can always reconsider that.

Reviewed-by: Chao Li <li.evan.chao@gmail.com>
Discussion: https://postgr.es/m/CA%2BhUKGKvr0x_oGmQTUkx%3DODgSksT2EtgCA6LmGx_jQFG%3DsDUpg%40mail.gmail.com