Drop support for MSVCRT's %I64 printf strings.
MSVCRT predated C99 and invented non-standard placeholders for 64-bit
numbers, and then later used them in standard macros when C99
<inttypes.h> arrived.  Those macros just use %lld etc when building with
UCRT, so there should be no way for our interposed sprintf.c code to
receive the pre-standard kind.  Time to drop the code that parses them.

That code was in fact already dead when commit 962da900 landed, as we'd
officially disclaimed MSVCRT support a couple of weeks earlier in commit
1758d424, but patch development overlapped and the history hadn't been
investigated.