aio: Simulate slow storage with I/O workers.
The following developer-only settings can be used to simulate the
characteristics of a system with limited I/O capacity.  They are
intended for studying system behavior under different conditions.  The
settings can be used alone or combined.

Fixed limit of IOs or bandwidth per second:

  debug_io_worker_limit_iops=1000
  debug_io_worker_limit_read='1MB'
  debug_io_worker_limit_write='500kB'

The submission queue size and behavior when it is full can also be
changed.  The default behavior is to fall back to synchronous I/O, but
if that is disabled with this setting then the behavior is to wait:

  debug_io_worker_queue_size=1024
  debug_io_worker_overflow_sync=false