Fix Makefile used for test_cloexec.
The previous test for ifeq ($(PORTNAME), win32) never succeeded, so only
meson-based builds were building and testing our O_CLOEXEC emulation.
Defect in commit c507ba55.

That could probably be made to work, but it seems better to compile
test_cloexec.c unconditionally and skip the test for non-Windows
systems.  Though the test is skipped on Unix and the executable wouldn't
test anything if run (there is not much point in testing that the system
O_CLOEXEC works on Unix), at least this way the configure-based build
machinery is consistent, for ease of maintenance by Unix-based
developers.

Backpatch-through: 16, like commit c507ba55